requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.6
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
idna==3.11
pycparser==3.0
PyGithub==2.9.0
PyJWT==2.12.1
PyNaCl==1.6.2
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
